requirements.txt
google-generative-ai
streamlit

